requirements.txt
torch
torchvision
facenet-pytorch
scikit-learn
Flask
Pillow
numpy
joblib
gunicorn
python-dotenv
matplotlib
pytest
